refactor(web): move to `web/src/tools/es-bundling`
Addresses code review comments.
