+++ web/build.sh
@@ -71,7 +71,7 @@ builder_describe_outputs \
   build:tools                   "/web/build/tools/building/sourcemap-root/index.js" \
   build:test-pages              "/web/build/test-resources/sentry-manager.js"
 
-BUNDLE_CMD="node ${KEYMAN_ROOT}/common/tools/es-bundling/build/common-bundle.mjs"
+BUNDLE_CMD="node ${KEYMAN_ROOT}/web/src/tools/es-bundling/build/common-bundle.mjs"
 
 #### Build action definitions ####
