+++ web/src/engine/keyboard/build.sh
@@ -17,10 +17,10 @@ SUBPROJECT_NAME=engine/keyboard
 
 builder_describe \
   "Compiles the web-oriented utility function module." \
-  "@/web/src/tools/testing/recorder-core  test" \
   "@/common/web/keyman-version" \
-  "@/common/tools/es-bundling" \
   "@/common/web/types" \
+  "@/web/src/tools/testing/recorder-core  test" \
+  "@/web/src/tools/es-bundling" \
   "@/web/src/engine/common/web-utils" \
   configure \
   clean \
