+++ web/src/engine/osk/gesture-processor/src/tools/unit-test-resources/build-bundler.js
@@ -1,5 +1,5 @@
 import esbuild from 'esbuild';
-import { esmConfiguration, bundleObjEntryPoints } from '../../../../es-bundling/build/index.mjs';
+import { esmConfiguration, bundleObjEntryPoints } from '../../../../../../tools/es-bundling/build/index.mjs';
 
 await esbuild.build({
   ...esmConfiguration,
