+++ web/src/engine/predictive-text/templates/build.sh
@@ -14,7 +14,7 @@ THIS_SCRIPT="$(readlink -f "${BASH_SOURCE[0]}")"
 
 builder_describe "Builds the predictive-text model template implementation module" \
   "@/common/web/keyman-version" \
-  "@/common/tools/es-bundling" \
+  "@/web/src/tools/es-bundling" \
   "@../wordbreakers" \
   "clean" \
   "configure" \
