+++ web/src/engine/predictive-text/worker-main/build.sh
@@ -13,14 +13,14 @@ THIS_SCRIPT="$(readlink -f "${BASH_SOURCE[0]}")"
 
 . "$KEYMAN_ROOT/resources/shellHelperFunctions.sh"
 
-BUNDLE_CMD="node $KEYMAN_ROOT/common/tools/es-bundling/build/common-bundle.mjs"
+BUNDLE_CMD="node $KEYMAN_ROOT/web/src/tools/es-bundling/build/common-bundle.mjs"
 
 ################################ Main script ################################
 
 builder_describe "Builds the lm-layer module" \
   "@/common/web/keyman-version" \
-  "@/common/tools/es-bundling" \
   "@/common/web/types" \
+  "@/web/src/tools/es-bundling" \
   "@/web/src/engine/predictive-text/worker-thread" \
   "clean" \
   "configure" \
