+++ web/src/engine/predictive-text/worker-thread/build.sh
@@ -17,7 +17,7 @@ WORKER_OUTPUT_FILENAME=build/lib/worker-main.js
 INTERMEDIATE=./build/intermediate
 LIB=./build/lib
 
-bundle_cmd="node ${KEYMAN_ROOT}/common/tools/es-bundling/build/common-bundle.mjs"
+bundle_cmd="node ${KEYMAN_ROOT}/web/src/tools/es-bundling/build/common-bundle.mjs"
 
 SRCMAP_CLEANER="node $KEYMAN_ROOT/web/build/tools/building/sourcemap-root/index.js"
 
@@ -25,9 +25,9 @@ SRCMAP_CLEANER="node $KEYMAN_ROOT/web/build/tools/building/sourcemap-root/index.
 
 builder_describe \
   "Compiles the Language Modeling Layer for common use in predictive text and autocorrective applications." \
-  "@/web/src/tools/building/sourcemap-root" \
   "@/common/web/keyman-version" \
-  "@/common/tools/es-bundling" \
+  "@/web/src/tools/building/sourcemap-root" \
+  "@/web/src/tools/es-bundling" \
   "@../wordbreakers" \
   "@../templates" \
   configure clean build test --ci
