+++ web/src/tools/es-bundling/build.sh
@@ -3,7 +3,7 @@
 ## START STANDARD BUILD SCRIPT INCLUDE
 # adjust relative paths as necessary
 THIS_SCRIPT="$(readlink -f "${BASH_SOURCE[0]}")"
-. "${THIS_SCRIPT%/*}/../../../resources/build/builder.inc.sh"
+. "${THIS_SCRIPT%/*}/../../../../resources/build/builder.inc.sh"
 ## END STANDARD BUILD SCRIPT INCLUDE
 
 . "$KEYMAN_ROOT/resources/shellHelperFunctions.sh"
@@ -17,7 +17,7 @@ builder_describe "Builds KMW's esbuild-oriented common configuration & tooling" 
 
 builder_describe_outputs \
   configure          /node_modules \
-  build              /common/tools/es-bundling/build/index.mjs
+  build              /web/src/tools/es-bundling/build/index.mjs
 
 builder_parse "$@"
