+++ web/src/tools/es-bundling/src/common-bundle.mts
