+++ web/src/tools/es-bundling/src/configuration.mts
